requirements.txt
passlib
mongoengine
pydantic
pymongo
h11
pip
wheel
cryptography
keyring
click
asgiref
starlette
certifi
uvicorn
idna
setuptools
bcrypt
python-jose
python-multipart
pytest
mongomock
python-decouple
pyjwt
email-validator
fastapi
requests
uvicorn
jinja2
urllib3
prettytable
streamlit
cloudscraper
redis
redmail
pandas
apscheduler
discord